Separate modules for network/model/trending movies
