+++ app/src/main/java/com/seungleekim/android/movie/model/Movie.kt
@@ -1,4 +1,4 @@
-package com.seungleekim.android.movie
+package com.seungleekim.android.movie.model
 
 import android.os.Parcelable
 import com.google.gson.annotations.SerializedName
