+++ app/src/main/java/com/seungleekim/android/movie/network/TmdbApi.kt
@@ -1,5 +1,6 @@
-package com.seungleekim.android.movie
+package com.seungleekim.android.movie.network
 
+import com.seungleekim.android.movie.network.response.TrendingMoviesResponse
 import io.reactivex.Observable
 import retrofit2.http.GET
