+++ app/src/main/java/com/seungleekim/android/movie/network/response/TrendingMoviesResponse.kt
@@ -1,6 +1,7 @@
-package com.seungleekim.android.movie
+package com.seungleekim.android.movie.network.response
 
 import com.google.gson.annotations.SerializedName
+import com.seungleekim.android.movie.model.Movie
 
 class TrendingMoviesResponse {
     @SerializedName("results")
