+++ app/src/main/res/layout/activity_trending.xml
@@ -5,4 +5,4 @@
     android:id="@+id/fragment_container"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".TrendingActivity" />
+    tools:context=".trending.TrendingMoviesActivity" />
