+++ app/src/main/res/layout/fragment_trending.xml
@@ -6,7 +6,7 @@
         android:id="@+id/trending_container"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".TrendingFragment">
+        tools:context=".trending.TrendingMoviesFragment">
 
     <android.support.design.widget.AppBarLayout
             android:id="@+id/app_bar"
